move fixtures to dedicated browser tests conftest file
